+++ tests/test_utils.sh
@@ -9,7 +9,7 @@ trap test_int INT
 function test_int() {
   echo "Interrupting test"
   docker compose ${COMPOSE_FLAGS} --project-directory ${PWD} --project-name ${PROJECT} down
-  rm -rf ${DBROOT}/{db,media,migrations}
+  rm -rf ${DBROOT:?}/{db,media,migrations}
   docker network rm ${PROJECT}-${NETWORK}
   [ -n "${COMPOSE_DELETE}" ] && rm -f ${COMPOSE_DELETE}
   echo "Test aborted"
@@ -22,7 +22,6 @@ function wait_for_container()
   WAITFORSTRING=${2:-"Container is ready"}
   MAX_WAIT=${3:-"60"}
   CUR_WAIT=0
-  CONTAINER_READY=0
   while [ -z "$(docker ps -q -f name=^/${CONTAINERNAME}$ )" ]
   do
     sleep 1
@@ -38,7 +37,6 @@ function wait_for_container()
   do
     if docker logs ${CONTAINERNAME} 2>&1 | grep -q "${WAITFORSTRING}"
     then
-      CONTAINER_READY=1
       break
     fi
     sleep 1
